comments intergrated with auth

+++ controllers/comment.js
@@ -2,6 +2,7 @@ const Comment = require("../models/comment");
 
 exports.createComment = async (req, res) => {
   try {
+    req.body.userId = req.user._id;
     const comment = await Comment.create(req.body);
 
     res.status(201).json({
@@ -10,7 +11,6 @@ exports.createComment = async (req, res) => {
         comment: comment,
       },
     });
-    
   } catch (err) {
     res.status(400).json({
       status: "failed",
@@ -19,7 +19,7 @@ exports.createComment = async (req, res) => {
   }
 };
 
-exports.getAllComments= async (req, res) => {
+exports.getAllComments = async (req, res) => {
   try {
     const comments = await Comment.find();
     console.log(comments);
@@ -31,7 +31,6 @@ exports.getAllComments= async (req, res) => {
         comments,
       },
     });
-
   } catch (err) {
     res.status(404).json({
       status: "fail",
@@ -50,7 +49,6 @@ exports.getComment = async (req, res) => {
         comment,
       },
     });
-
   } catch (err) {
     res.status(404).json({
       status: "fail",
@@ -61,19 +59,33 @@ exports.getComment = async (req, res) => {
 
 exports.updateComment = async (req, res) => {
   try {
-    const comment = await Comment.findByIdAndUpdate(req.params.id, req.body, {
-      new: true,
-      runValidators: true,
-    });
+    const comment = await Comment.findById(req.params.id);
+    if (!comment) {
+      res.status(400);
+      throw new Error("Please Enter a Valid comment");
+    }
+    if (comment.userId.toString() !== req.user._id.toString()) {
+      // if (comment.userId !== req.user._id) {
+      res.status(400);
+      throw new Error("Please add a text field");
+    }
 
+    const updatedComment = await Comment.findByIdAndUpdate(
+      req.params.id,
+      req.body,
+      {
+        new: true,
+        runValidators: true,
+      }
+    );
     res.status(200).json({
       status: "success",
       data: {
-        comment,
+        updatedComment,
       },
     });
-
   } catch (err) {
+    console.log(err);
     res.status(404).json({
       status: "fail",
       message: err,
@@ -83,6 +95,19 @@ exports.updateComment = async (req, res) => {
 
 exports.deleteComment = async (req, res) => {
   try {
+    const comment = await Comment.findById(req.params.id);
+    if (!comment) {
+      res.status(400);
+      throw new Error("Please Enter a Valid comment");
+    }
+    console.log(comment.userId)
+    console.log(`req.user._id = ${req.user._id}`)
+    if (comment.userId.toString() !== req.user._id.toString()) {
+      // if (comment.userId !== req.user._id) {
+      res.status(400);
+      throw new Error("Please add a text field");
+    }
+
     await Comment.findByIdAndDelete(req.params.id);
 
     res.status(204).json({
@@ -90,8 +115,8 @@ exports.deleteComment = async (req, res) => {
       message: "comment deleted successfully",
       data: null,
     });
-
   } catch (err) {
+    console.log(err)
     res.status(404).json({
       status: "fail",
       message: err,
